requirements.txt
google-generativeai
requests
pandas
scikit-learn
python-dotenv
